pyproject.toml
[tool.poetry]
name = "my-course"
version = "0.1.0"
description = ""
authors = ["nura"]
readme = "README.md"
packages = [{include = "my_course"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.93.0"
uvicorn = "^0.20.0"
fastapi-contrib = "^0.2.11"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.5.post1"
alembic = "^1.10.2"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
sqlalchemy-utils = "^0.40.0"
pytz = "^2022.7.1"
pymongo = "^4.3.3"
motor = "^3.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
